requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
dialogflow==1.1.1
feedparser==6.0.8
Flask==2.0.1
fuzzywuzzy==0.18.0
gnewsclient==1.12
google-api-core==1.31.2
google-auth==1.35.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==12.8
pytz==2021.1
requests==2.26.0
rsa==4.7.2
sgmllib3k==1.0.0
six==1.16.0
tornado==6.1
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
Werkzeug==2.0.1
